+++ lrucore/core.go
@@ -8,9 +8,6 @@ import (
 	"errors"
 	"math"
 	"sync"
-	"time"
-
-	"github.com/justpranavrs/tlru/lruclock"
 )
 
 // Core is basic implementation of 'Least Recently Used' Cache.
@@ -43,21 +40,6 @@ type Core[K comparable, V any] struct {
 
 	// capacity represents the maximum allocated space for the cache.
 	capacity int
-
-	// hasTTL is initialized when the instance is created.
-	// If true, ttl is enabled. Defaults to false.
-	hasTTL bool
-
-	// clock is the background timer to ensure fast time loads
-	// without halting the LRU operations.
-	clock *lruclock.Clock
-
-	// expiresAt determines the (time-to-live) duration of an element.
-	expiresAt int64
-
-	// timestamps stores the last updated timestamp of a doubly-linked list
-	// element and is set during [Core.addKey] and updated at [Core.putKey].
-	timestamps []int64
 
 	// stats measures the instance's hits, misses and evictions.
 	stats CoreStats
@@ -120,46 +102,22 @@ const (
 )
 
 var (
-	// ErrInvalidBatchSize is returned by [Core.GetMany] or [Core.PutMany] when keys and values do not have the same lengths.
+	// ErrInvalidBatchSize is returned by batch operations when keys and values do not have the same lengths.
 	ErrInvalidBatchSize = errors.New("invalid LRU batch sizes: keys and values do not have the same lengths")
 
-	// ErrInvalidCapacity is returned by [New] when the maximum cache capacity is not in [2, 2147483646].
+	// ErrInvalidCapacity is returned by [New] or [NewTTL] when the maximum cache capacity is not in [2, 2147483646].
 	ErrInvalidCapacity = errors.New("invalid LRU cache capacity: must be in the range [2, 2147483646]")
 )
-
-// coreConfig represents the configuration of [Core]. It should be used with [CoreOption].
-type coreConfig struct {
-	hasTTL    bool
-	expiresAt time.Duration
-
-	clock *lruclock.Clock
-}
-
-// CoreOption is used to configure [Core] when creating an instance using [New] constructor.
-type CoreOption func(c *coreConfig) error
 
 // New creates an instance of [Core] using the given capacity.
 //
 // Returns an [ErrInvalidCapacity] if the capacity is not in [2, 2147483646].
-func New[K comparable, V any](capacity int, opts ...CoreOption) (*Core[K, V], error) {
+func New[K comparable, V any](capacity int) (*Core[K, V], error) {
 	if capacity < 2 || (capacity > math.MaxInt32-2) { // For capacity 1, a variable can be used.
 		return nil, ErrInvalidCapacity
 	}
 	cap := int32(capacity)
 	tail := 1 + cap
-
-	// build the config
-	cfg := coreConfig{
-		hasTTL: false,
-	}
-	for _, opt := range opts { // options
-		if opt == nil {
-			continue
-		}
-		if err := opt(&cfg); err != nil {
-			return nil, err
-		}
-	}
 
 	// allocate the initial nodes and links array with a size of 2+cap
 	nodes := make([]node[K, V], 2+capacity)
@@ -175,7 +133,7 @@ func New[K comparable, V any](capacity int, opts ...CoreOption) (*Core[K, V], er
 	}
 
 	// lru
-	lru := &Core[K, V]{
+	return &Core[K, V]{
 		hash:     make(map[K]int32, capacity),
 		nodes:    nodes,
 		links:    links,
@@ -188,52 +146,7 @@ func New[K comparable, V any](capacity int, opts ...CoreOption) (*Core[K, V], er
 			Evictions:   0,
 			Expirations: 0,
 		},
-		hasTTL:     cfg.hasTTL,
-		timestamps: nil,
-		clock:      nil,
-	}
-
-	if cfg.hasTTL {
-		lru.timestamps = make([]int64, 2+capacity)
-
-		duration := 100 * time.Millisecond
-		if cfg.clock != nil {
-			lru.clock = cfg.clock
-			duration = cfg.clock.Duration()
-		} else {
-			lru.clock = lruclock.New(duration)
-			_ = lru.clock.Start()
-		}
-		lru.expiresAt = int64(cfg.expiresAt / duration)
-	}
-	return lru, nil
-}
-
-// WithClock allows the usage of a custom clock for [Core].
-// It is only initialized if "TTL" is enabled.
-//
-// NOTE: Using WithClock on [New] will not start the clock. Use [lruclock.Clock.Start] to
-// initiate the timer.
-func WithClock(clock *lruclock.Clock) CoreOption {
-	return func(c *coreConfig) error {
-		c.clock = clock
-		return nil
-	}
-}
-
-// WithTTL enables (Time-to-Live) for [Core].
-// expiresAt determines the (TTL) duration of an element in the cache.
-//
-// [Core] uses Absolute TTL if enabled. It uses a clock with a duration of 100ms.
-// To customize the clock, check [WithClock].
-//
-// Enabling TTL creates a Background clock. To close it use [Core.Close].
-func WithTTL(expiresAt time.Duration) CoreOption {
-	return func(c *coreConfig) error {
-		c.hasTTL = true
-		c.expiresAt = expiresAt
-		return nil
-	}
+	}, nil
 }
 
 // Capacity returns the maximum allocated capacity of the LRU cache.
@@ -241,15 +154,10 @@ func (l *Core[K, V]) Capacity() int {
 	return l.capacity
 }
 
-// Close safely closes the background clock when TTL is enabled on the cache.
-func (l *Core[K, V]) Close() {
-	l.mu.Lock()
-	defer l.mu.Unlock()
-
-	if l.clock != nil {
-		l.clock.Stop()
-	}
-}
+// Close safely closes the background clock on the cache.
+//
+// NOTE: [Core] has no internal clock to safely shutdown. This does nothing.
+func (l *Core[K, V]) Close() {}
 
 // Delete removes the key from the cache and returns the evicted value.
 // It returns false if the key was not found in the cache.
@@ -290,7 +198,7 @@ func (l *Core[K, V]) Get(key K) (V, bool) {
 // [ErrInvalidBatchSize] is returned.
 //
 // The operation modifies values and exists in-place. If a key is not present in the cache,
-// the corresponding index in exists tis set to false and leaves the value at that index unchanged.
+// the corresponding index in exists is set to false and leaves the value at that index unchanged.
 func (l *Core[K, V]) GetMany(keys []K, values []V, exists []bool) error {
 	if (len(keys) != len(values)) || (len(keys) != len(exists)) {
 		return ErrInvalidBatchSize
@@ -354,7 +262,7 @@ func (l *Core[K, V]) ResetStats() {
 }
 
 // Shards returns the number of sharded instances in the LRU cache.
-// For [lrucore.Core], this is always 1.
+// For [Core], this is always 1.
 func (l *Core[K, V]) Shards() int {
 	return 1
 }
@@ -376,8 +284,13 @@ func (l *Core[K, V]) Stats() CoreStats {
 }
 
 // Upsert adds a new value to the cache with the given key.
-// It returns a value based on how the internal state of the cache changed.
-func (l *Core[K, V]) Upsert(key K, value V) UpsertState {
+// It returns [UpsertState] based on how the internal state of the cache changed.
+//
+// It also returns a value based on [UpsertState]
+//   - [AddNoEvict] returns nothing.
+//   - [AddOnEvict] returns the evicted value.
+//   - [Replace] returns the old value the key had.
+func (l *Core[K, V]) Upsert(key K, value V) (UpsertState, V) {
 	l.mu.Lock()
 	defer l.mu.Unlock()
 
@@ -391,12 +304,19 @@ func (l *Core[K, V]) addKey(key K, value V) {
 		key:   key,
 		value: value,
 	}
-	if l.hasTTL {
-		l.timestamps[free] = l.clock.Now()
-	}
 
 	l.hash[key] = free
 	l.mru = free
+}
+
+// getIndex updates the recency, the stats and returns value with the index.
+func (l *Core[K, V]) getIndex(idx int32) V {
+	if idx != l.mru { // not already recent
+		l.makeRecent(idx)
+	}
+	l.stats.Hits++
+	val := l.nodes[idx].value
+	return val
 }
 
 // getKey retrieves the value of the key passed as argument.
@@ -407,27 +327,14 @@ func (l *Core[K, V]) getKey(key K) (V, bool) {
 		l.stats.Misses++
 		return *new(V), false // not present in cache
 	}
-	if l.hasTTL && l.clock.Since(l.timestamps[curr]) > l.expiresAt {
-		l.deleteKey(curr)
-		l.stats.Expirations++
-
-		l.stats.Misses++
-		return *new(V), false
-	}
-
-	if curr != l.mru { // not already recent
-		l.remove(curr)
-		l.makeRecent(curr)
-	}
-	l.stats.Hits++
-	val := l.nodes[curr].value
-	return val, true
+	return l.getIndex(curr), true
 }
 
 // makeRecent sets the key as 'Most Recently Used'.
 func (l *Core[K, V]) makeRecent(idx int32) {
-	free := l.links[l.mru].next
+	l.remove(idx)
 
+	free := l.links[l.mru].next
 	l.links[l.mru].next = idx
 	l.links[free].prev = idx
 
@@ -445,46 +352,35 @@ func (l *Core[K, V]) peekKey(key K) (V, bool) {
 	if !ok {
 		return *new(V), false
 	}
-	if l.hasTTL && l.clock.Since(l.timestamps[curr]) > l.expiresAt {
-		l.deleteKey(curr)
-		l.stats.Expirations++
-
-		l.stats.Misses++
-		return *new(V), false
-	}
 	return l.nodes[curr].value, true
 }
 
 // putKey adds or updates the cache value with key "key".
-func (l *Core[K, V]) putKey(key K, value V) UpsertState {
+// It also returns a value based on whether it was evicted or replaced.
+func (l *Core[K, V]) putKey(key K, value V) (UpsertState, V) {
 	curr, ok := l.hash[key]
 	if !ok { // not present in cache
 		if len(l.hash) == l.capacity {
 			old := l.links[0].next
+			oldVal := l.nodes[old].value
+
 			l.deleteKey(old)
 			l.stats.Evictions++
 
 			l.addKey(key, value)
-			return AddOnEvict
+			return AddOnEvict, oldVal
 		} else {
 			l.addKey(key, value)
-			return AddNoEvict
+			return AddNoEvict, *new(V)
 		}
 	} else { // present in cache, just update values
+		oldVal := l.nodes[curr].value
+
 		l.nodes[curr].value = value
 		if curr != l.mru { // not already recent
-			l.remove(curr)
 			l.makeRecent(curr)
 		}
-		if l.hasTTL {
-			if l.clock.Since(l.timestamps[curr]) > l.expiresAt {
-				l.stats.Expirations++
-				l.timestamps[curr] = l.clock.Now()
-				return AddAfterExpiration
-			}
-			l.timestamps[curr] = l.clock.Now()
-		}
-		return Replace
+		return Replace, oldVal
 	}
 }